authentication with auth0 integrated
