--- meow/auth_config.js
@@ -1,8 +1,0 @@
-// auth_config.js
-
-const config = {
-    domain: "dev-ayqodrm6v6my2o1d.us.auth0.com",
-    clientId: "zkPWjCKqKslsBL6K53DTy9325ZRlglYv"
-  };
-  
-  export default config;
