+++ meow/package.json
@@ -32,7 +32,7 @@
 	},
 	"type": "module",
 	"dependencies": {
-		"@auth0/auth0-spa-js": "^2.0.3",
+		"@auth0/auth0-spa-js": "^2.0.4",
 		"@types/node": "^18.13.0",
 		"autoprefixer": "^10.4.13",
 		"cors": "^2.8.5",
