--- meow/src/components/TaskItem.svelte
@@ -1,44 +1,0 @@
-<script>
-    import { tasks } from "../store";
-    export let task = {};
-  
-    let isChecked;
-  
-    function taskDone() {
-      console.log(isChecked);
-  
-      let updatedTasks = $tasks.map((currentTask) => {
-        if (currentTask.id === task.id) {
-          currentTask.completed = isChecked;
-          return currentTask;
-        }
-        return currentTask;
-      });
-  
-      tasks.set(updatedTasks);
-      console.log($tasks);
-    }
-  </script>
-  
-  <style>
-    .completed {
-      color: red;
-      text-decoration: line-through;
-    }
-  </style>
-  
-  <main>
-    <li class="list-group-item">
-      <input
-        type="checkbox"
-        class="form-check-input"
-        id="exampleCheck1"
-        bind:checked={isChecked}
-        on:change={(e) => 
-        taskDone() 
-        } />
-      <span class:completed={task.completed}>{task.description}</span>
-    </li>
-  </main>
-
-  //note taskDone(e) on 38 https://stackoverflow.com/questions/62146182/typescript-is-giving-me-an-error-expected-0-arguments-but-got-1
