+++ meow/src/lib/config/auth_config.js
@@ -1,0 +1,8 @@
+// auth_config.js
+
+const config = {
+    domain: "dev-ayqodrm6v6my2o1d.us.auth0.com",
+    clientId: "zkPWjCKqKslsBL6K53DTy9325ZRlglYv"
+  };
+  
+export default config;
