+++ meow/src/lib/services/authService.js
@@ -1,0 +1,41 @@
+// src/authService.js
+
+import { createAuth0Client } from "@auth0/auth0-spa-js";
+import { user, isAuthenticated, popupOpen } from "../stores/store";
+import config from "../config/auth_config";
+
+async function createClient() {
+  let auth0Client = await createAuth0Client({
+    domain: config.domain,
+    clientId: config.clientId
+  });
+
+  return auth0Client;
+}
+
+async function loginWithPopup(client, options) {
+  popupOpen.set(true);
+  try {
+    await client.loginWithPopup(options);
+
+    user.set(await client.getUser());
+    isAuthenticated.set(true);
+  } catch (e) {
+    // eslint-disable-next-line
+    console.error(e);
+  } finally {
+    popupOpen.set(false);
+  }
+}
+
+function logout(client) {
+  return client.logout();
+}
+
+const auth = {
+  createClient,
+  loginWithPopup,
+  logout
+};
+
+export default auth;
