+++ meow/src/routes/Header.svelte
@@ -3,25 +3,24 @@
 	import logo from '$lib/images/meow.png';
 	import profile from '$lib/images/profile.png';
 	import { Dropdown, DropdownItem } from 'flowbite-svelte'
+	import auth from '$lib/services/authService'
+	import { isAuthenticated, user } from '$lib/stores/store'
+	import { onMount } from 'svelte'
+	
+	let auth0Client = {}
+	
+	onMount(async () => {
+		auth0Client = await auth.createClient()
+		isAuthenticated.set(await auth0Client.isAuthenticated())
+		user.set(await auth0Client.getUser())
+	})
 
-	function logOut() {
-		var x = document.getElementById("login");
-		var y = document.getElementById("signup");
-		var z = document.getElementById("profile-icon");
-		
-		x.className = "btn-primary";
-		y.className = "btn-primary";
-		z.style.display = "none";
+	function login() {
+		auth.loginWithPopup(auth0Client)
 	}
 
-	function userReturn() {
-		var x = document.getElementById("login");
-		var y = document.getElementById("signup");
-		var z = document.getElementById("profile-icon");
-		
-		x.className = "btn-invis";
-		y.className = "btn-invis";
-		z.style.display = "block";
+	function logout() {
+		auth.logout(auth0Client)
 	}
 
 </script>
@@ -66,25 +65,42 @@
 		</svg>
 	</nav>
 
+	{#if !$isAuthenticated}
 	<div class="flex justify-between">
 		<div class="p-2 flex space-x-2 justify-center object-contain">
-			<button id="signup" type="button" class="btn-primary" on:click={userReturn}>Sign Up</button>
+			<button id="signup" type="button" class="btn-primary" on:click={login}>Sign Up</button>
 		</div>
 		<div class="p-2 flex space-x-2 justify-center object-contain">
-			<button id="login" type="button" class="btn-primary" on:click={userReturn}>Log In</button>
+			<button id="login" type="button" class="btn-primary" on:click={login}>Log In</button>
 		</div>
 		
 		<div class="w-12 h-12">
 			<button id="profile-icon" style="display:none" class="flex items-center justify-center w-full h-full">
 				<img src={profile} alt="Profile" class="w-8 h-8 object-contain"/>
 			</button>
+		</div>
+	</div>
+	{:else}
+	<div class="flex justify-between">
+		<div class="p-2">
+			<button type="button" class="btn-invis" disabled>Sign Up</button>
+		</div>
+		<div class="p-2">
+			<button type="button" class="btn-invis" disabled>Log In</button>
+		</div>
+	
+		<div class="w-12 h-12">
+			<button id="profile-icon" class="flex items-center justify-center w-full h-full">
+				<img src={profile} alt="Profile" class="w-8 h-8 object-contain"/>
+			</button>
 			<Dropdown>
 				<DropdownItem>Profile</DropdownItem>
 				<DropdownItem>Settings</DropdownItem>
-				<DropdownItem on:click={logOut}>Log Out</DropdownItem>
+				<DropdownItem on:click={logout}>Log Out</DropdownItem>
 			</Dropdown>
 		</div>
 	</div>
+	{/if}	
 </header>
 
 <style>
