--- meow/src/store.js
@@ -1,31 +1,0 @@
-// src/store.js
-
-import { writable, derived, readable } from "svelte/store";
-
-
-// export const isAuthenticated = writable(false);
-// export const user = writable({});
-// export const popupOpen = writable(false);
-// export const error = writable();
-
-// export const tasks = writable([]);
-
-// export const user_tasks = derived([tasks, user], ([$tasks, $user]) => {
-//   let logged_in_user_tasks = [];
-
-//   if ($user && $user.email) {
-//     logged_in_user_tasks = $tasks.filter((task) => task.user === $user.email);
-//   }
-
-//   return logged_in_user_tasks;
-// });
-
-export const time = readable(new Date(), function start(set) {
-	const interval = setInterval(() => {
-		set(new Date());
-	}, 1000);
-
-	return function stop() {
-		clearInterval(interval);
-	};
-});
